Add files via upload
This data analysis project aims to analyze the sales performance of a company selling rice over the course of five years. The company sells rice at a selling price of 30 Rs per kg, with a buying price of 10 Rs per kg. In the first year, the company sells 2000 tons of rice. Additionally, there are Maintenance Charges (including storage and labor) and Delivery Charges (including transportation and labor) that are calculated as percentages of the buying price and reduce from the total profit. Historical data indicates an 11% increment in the buying price annually, with a corresponding increase in the selling price. Furthermore, there is an annual increment of 0.5% in Maintenance Charges and 0.8% in Delivery Charges. Considering a tax rate of 4%, this analysis aims to determine the revenue the company can generate by selling 3000 tons in the second year, 2500 tons in the third year, 3100 tons in the fourth year, and 1800 tons in the fifth year. The project seeks to identify trends, make data-driven recommendations, and gain insights into the company's sales performance.